Towards a coordinate grid based on the intersection points.
Thinking of the intersection points as the basis of a deformed coordinate grid will allow me to morph cartesian geometries onto the curved torus-like surface of the rosette.

This commit defines two dimensions: angle relative to the center of the guide circle, and distance from the center of the guide-circle.

Added the ability to render the resulting radial lines.